refactor: split __init__.py and test_base.py into focused submodules
- Extract JSONEncoder/JSONEncodeWarning/_UNENCODABLE_PLACEHOLDER_VALUE
  into microbench/_encoding.py
- Extract Output/FileOutput/RedisOutput into microbench/_output.py
- Extract all MB* mixin classes into microbench/mixins.py
- __init__.py now contains only MicroBench core + re-exports
- Rename MonitorThread -> _MonitorThread (internal-only, not in __all__)
- Split tests/test_base.py into test_core.py, test_mixins.py, test_output.py
- Public API (__all__) and all import paths unchanged
