--- microbench/tests/test_base.py
@@ -1,1730 +1,0 @@
-import atexit as _atexit
-import datetime
-import io
-import os
-import signal as _signal
-import sys
-import tempfile
-import threading
-import time
-import warnings
-from unittest.mock import MagicMock, patch
-
-import numpy
-import pandas
-import pytest
-
-from microbench import (
-    _UNENCODABLE_PLACEHOLDER_VALUE,
-    FileOutput,
-    JSONEncoder,
-    JSONEncodeWarning,
-    MBFileHash,
-    MBFunctionCall,
-    MBGitInfo,
-    MBHostInfo,
-    MBInstalledPackages,
-    MBLoadedModules,
-    MBPeakMemory,
-    MBPythonVersion,
-    MBReturnValue,
-    MBSlurmInfo,
-    MicroBench,
-    Output,
-    RedisOutput,
-)
-from microbench import __version__ as microbench_version
-
-from .globals_capture import globals_bench
-
-
-def test_mb_run_id_and_version():
-    """Every record contains mb_run_id (UUID) and mb_version."""
-    import re
-
-    import microbench
-
-    bench = MicroBench()
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-    noop()
-
-    results = bench.get_results()
-
-    # mb_run_id is a valid UUID and consistent across calls
-    uuid_re = re.compile(
-        r'^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
-    )
-    assert results['mb_run_id'].nunique() == 1
-    assert uuid_re.match(results['mb_run_id'][0])
-
-    # mb_version matches the installed package version
-    assert (results['mb_version'] == microbench.__version__).all()
-
-
-def test_mb_run_id_shared_across_instances():
-    """All MicroBench instances in the same process share the same mb_run_id."""
-    bench_a = MicroBench()
-    bench_b = MicroBench()
-
-    @bench_a
-    def func_a():
-        pass
-
-    @bench_b
-    def func_b():
-        pass
-
-    func_a()
-    func_b()
-
-    run_id_a = bench_a.get_results()['mb_run_id'][0]
-    run_id_b = bench_b.get_results()['mb_run_id'][0]
-    assert run_id_a == run_id_b
-
-
-def test_function():
-    class MyBench(MicroBench, MBFunctionCall, MBPythonVersion, MBHostInfo):
-        capture_versions = (pandas, io)
-        env_vars = ('TEST_NON_EXISTENT', 'HOME')
-
-    benchmark = MyBench(some_info='123')
-
-    @benchmark
-    def my_function():
-        """Inefficient function for testing"""
-        acc = 0
-        for i in range(1000000):
-            acc += i
-
-        return acc
-
-    for _ in range(3):
-        assert my_function() == 499999500000
-
-    results = benchmark.get_results()
-    assert (results['function_name'] == 'my_function').all()
-    assert results['package_versions'][0]['pandas'] == pandas.__version__
-    runtimes = results['finish_time'] - results['start_time']
-    assert (runtimes > datetime.timedelta(0)).all()
-
-    assert results['timestamp_tz'][0] == 'UTC'
-    assert results['duration_counter'][0] == 'perf_counter'
-
-
-def test_multi_iterations():
-    class MyBench(MicroBench):
-        pass
-
-    tz = datetime.timezone(datetime.timedelta(hours=10))
-    iterations = 3
-    benchmark = MyBench(iterations=iterations, tz=tz)
-
-    @benchmark
-    def my_function():
-        pass
-
-    # call the function
-    my_function()
-
-    results = benchmark.get_results()
-    assert (results['function_name'] == 'my_function').all()
-    runtimes = results['finish_time'] - results['start_time']
-    assert (runtimes >= datetime.timedelta(0)).all()
-    assert results['timestamp_tz'][0] == str(tz)
-    # Verify the timezone is actually applied to the timestamps, not just recorded
-    assert results['start_time'][0].utcoffset() == datetime.timedelta(hours=10)
-    assert results['finish_time'][0].utcoffset() == datetime.timedelta(hours=10)
-
-    assert len(results['run_durations'][0]) == iterations
-    assert all(dur >= 0 for dur in results['run_durations'][0])
-
-
-def test_capture_optional_records_errors():
-    """capture_optional=True catches failing captures, records in mb_capture_errors."""
-
-    class BrokenCapture(MicroBench):
-        capture_optional = True
-
-        def capture_will_fail(self, bm_data):
-            raise RuntimeError('simulated failure')
-
-    bench = BrokenCapture()
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-
-    results = bench.get_results()
-    errors = results['mb_capture_errors'][0]
-    assert len(errors) == 1
-    assert errors[0]['method'] == 'capture_will_fail'
-    assert 'RuntimeError' in errors[0]['error']
-    assert 'simulated failure' in errors[0]['error']
-
-
-def test_capture_optional_no_errors_no_field():
-    """When no captures fail, mb_capture_errors is absent from the record."""
-
-    class Bench(MicroBench):
-        capture_optional = True
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-
-    results = bench.get_results()
-    assert 'mb_capture_errors' not in results.columns
-
-
-def test_capture_optional_false_raises():
-    """Without capture_optional, a failing capture propagates the exception."""
-
-    class BrokenCapture(MicroBench):
-        def capture_will_fail(self, bm_data):
-            raise RuntimeError('simulated failure')
-
-    bench = BrokenCapture()
-
-    @bench
-    def noop():
-        pass
-
-    with pytest.raises(RuntimeError, match='simulated failure'):
-        noop()
-
-
-def test_capture_optional_capturepost():
-    """capture_optional also protects capturepost_ methods."""
-
-    class BrokenPost(MicroBench):
-        capture_optional = True
-
-        def capturepost_will_fail(self, bm_data):
-            raise ValueError('post failure')
-
-    bench = BrokenPost()
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-
-    results = bench.get_results()
-    errors = results['mb_capture_errors'][0]
-    assert any(e['method'] == 'capturepost_will_fail' for e in errors)
-
-
-def test_mb_slurm_info():
-    class Bench(MicroBench, MBSlurmInfo):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    slurm_env = {
-        'SLURM_JOB_ID': '12345',
-        'SLURM_ARRAY_TASK_ID': '3',
-        'SLURM_NODELIST': 'gpu-node-01',
-        'NOT_SLURM': 'ignored',
-    }
-
-    with patch.dict(os.environ, slurm_env, clear=False):
-        noop()
-
-    results = bench.get_results()
-    slurm = results['slurm'][0]
-    assert slurm['job_id'] == '12345'
-    assert slurm['array_task_id'] == '3'
-    assert slurm['nodelist'] == 'gpu-node-01'
-    assert 'not_slurm' not in slurm
-
-
-def test_mb_slurm_info_empty():
-    """slurm field is an empty dict when no SLURM_* vars are set."""
-
-    class Bench(MicroBench, MBSlurmInfo):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    # Strip any real SLURM vars that might be in the test environment
-    clean_env = {k: v for k, v in os.environ.items() if not k.startswith('SLURM_')}
-    with patch.dict(os.environ, clean_env, clear=True):
-        noop()
-
-    results = bench.get_results()
-    assert results['slurm'][0] == {}
-
-
-def test_mb_loaded_modules():
-    """loaded_modules captures name/version pairs from LOADEDMODULES."""
-
-    class Bench(MicroBench, MBLoadedModules):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch.dict(
-        os.environ,
-        {'LOADEDMODULES': 'gcc/12.2.0:openmpi/4.1.5:python/3.10.4'},
-        clear=False,
-    ):
-        noop()
-
-    modules = bench.get_results()['loaded_modules'][0]
-    assert modules['gcc'] == '12.2.0'
-    assert modules['openmpi'] == '4.1.5'
-    assert modules['python'] == '3.10.4'
-
-
-def test_mb_loaded_modules_empty():
-    """loaded_modules is an empty dict when LOADEDMODULES is unset."""
-
-    class Bench(MicroBench, MBLoadedModules):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    clean_env = {k: v for k, v in os.environ.items() if k != 'LOADEDMODULES'}
-    with patch.dict(os.environ, clean_env, clear=True):
-        noop()
-
-    assert bench.get_results()['loaded_modules'][0] == {}
-
-
-def test_mb_loaded_modules_no_version():
-    """Module entries without a version store an empty string."""
-
-    class Bench(MicroBench, MBLoadedModules):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch.dict(os.environ, {'LOADEDMODULES': 'null:gcc/12.2.0'}, clear=False):
-        noop()
-
-    modules = bench.get_results()['loaded_modules'][0]
-    assert modules['null'] == ''
-    assert modules['gcc'] == '12.2.0'
-
-
-def test_mb_loaded_modules_version_with_slash():
-    """Module versions containing slashes are captured in full."""
-
-    class Bench(MicroBench, MBLoadedModules):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    # Some module systems use hierarchical names like GCC/12.2.0-GCCcore-12.2.0
-    with patch.dict(
-        os.environ,
-        {'LOADEDMODULES': 'GCC/12.2.0-GCCcore-12.2.0'},
-        clear=False,
-    ):
-        noop()
-
-    modules = bench.get_results()['loaded_modules'][0]
-    assert modules['GCC'] == '12.2.0-GCCcore-12.2.0'
-
-
-def test_warmup():
-    call_count = 0
-
-    bench = MicroBench(warmup=3, iterations=2)
-
-    @bench
-    def my_function():
-        nonlocal call_count
-        call_count += 1
-
-    my_function()
-
-    # 3 warmup + 2 recorded iterations = 5 total calls
-    assert call_count == 5
-
-    results = bench.get_results()
-    # Only one record (one decorated call), with 2 run_durations
-    assert len(results) == 1
-    assert len(results['run_durations'][0]) == 2
-
-
-def test_local_timezone():
-    """Verify README example syntax: tz=datetime.datetime.now().astimezone().tzinfo.
-
-    This is a smoke test that the expression produces a valid timezone accepted
-    by MicroBench, and that the stored offset matches whatever was passed in.
-    On UTC machines the offset is timedelta(0) — identical to the default — so
-    this test does not discriminate between 'tz= applied' and 'tz= ignored'.
-    test_multi_iterations covers the non-UTC case with a hardcoded UTC+10 offset.
-    """
-
-    class MyBench(MicroBench):
-        pass
-
-    local_tz = datetime.datetime.now().astimezone().tzinfo
-    benchmark = MyBench(tz=local_tz)
-
-    @benchmark
-    def noop():
-        pass
-
-    noop()
-
-    results = benchmark.get_results()
-    expected_offset = datetime.datetime.now().astimezone().utcoffset()
-    assert results['start_time'][0].utcoffset() == expected_offset
-    assert results['finish_time'][0].utcoffset() == expected_offset
-    assert results['timestamp_tz'][0] == str(local_tz)
-
-
-def test_capture_global_packages():
-    @globals_bench
-    def noop():
-        pass
-
-    noop()
-
-    results = globals_bench.get_results()
-
-    # We should've captured microbench and pandas versions from top level
-    # imports in this file
-    assert results['package_versions'][0]['microbench'] == str(microbench_version)
-    assert results['package_versions'][0]['pandas'] == pandas.__version__
-
-
-def test_capture_packages_importlib():
-    class PkgBench(MicroBench, MBInstalledPackages):
-        capture_paths = True
-
-    pkg_bench = PkgBench()
-
-    @pkg_bench
-    def noop():
-        pass
-
-    noop()
-
-    results = pkg_bench.get_results()
-    assert pandas.__version__ == results['package_versions'][0]['pandas']
-
-
-def test_capture_packages_self_imports_metadata():
-    """capture_packages imports importlib.metadata itself, not via prior imports."""
-    import sys
-
-    # Evict importlib.metadata so the method's own import statement is exercised.
-    # Without 'import importlib.metadata' inside capture_packages, this would raise
-    # AttributeError: module 'importlib' has no attribute 'metadata'.
-    saved = sys.modules.pop('importlib.metadata', None)
-    try:
-
-        class PkgBench(MicroBench, MBInstalledPackages):
-            pass
-
-        bench = PkgBench()
-
-        @bench
-        def noop():
-            pass
-
-        noop()
-
-        results = bench.get_results()
-        assert isinstance(results['package_versions'][0], dict)
-        assert len(results['package_versions'][0]) > 0
-    finally:
-        if saved is not None:
-            sys.modules['importlib.metadata'] = saved
-
-
-def test_monitor():
-    class MonitorBench(MicroBench):
-        @staticmethod
-        def monitor(process):
-            return process.memory_full_info()._asdict()
-
-    monitor_bench = MonitorBench()
-
-    @monitor_bench
-    def noop():
-        pass
-
-    noop()
-
-    # Check monitor thread completed
-    assert not monitor_bench._monitor_thread.is_alive()
-
-    # Check some monitor data was captured
-    results = monitor_bench.get_results()
-    assert len(results['monitor']) > 0
-
-
-def test_monitor_from_non_main_thread():
-    """Monitor must not crash when started from a non-main thread (B6 fix).
-
-    signal.signal() can only be called from the main thread; MonitorThread
-    should skip signal registration and emit a RuntimeWarning instead.
-    """
-
-    class MonitorBench(MicroBench):
-        @staticmethod
-        def monitor(process):
-            return {'rss': process.memory_info().rss}
-
-    monitor_bench = MonitorBench()
-
-    @monitor_bench
-    def noop():
-        pass
-
-    errors = []
-    caught_warnings = []
-
-    def run_from_thread():
-        with warnings.catch_warnings(record=True) as w:
-            warnings.simplefilter('always')
-            try:
-                noop()
-            except Exception as exc:
-                errors.append(exc)
-            caught_warnings.extend(w)
-
-    t = threading.Thread(target=run_from_thread)
-    t.start()
-    t.join()
-
-    assert not errors, f'Unexpected exception from non-main thread: {errors[0]}'
-    assert any(issubclass(w.category, RuntimeWarning) for w in caught_warnings), (
-        'Expected a RuntimeWarning about signal registration being skipped'
-    )
-
-
-def test_functioncall_args_not_double_encoded():
-    """MBFunctionCall must store raw values, not JSON strings (B5 fix).
-
-    Before the fix, self.to_json(v) stored a JSON string which then got
-    re-serialized, turning e.g. {'k': 1} into the string '{"k": 1}'.
-    """
-
-    class Bench(MicroBench, MBFunctionCall):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def dummy(pos_int, pos_dict, kw_str='default'):
-        pass
-
-    dummy(42, {'key': 'value'}, kw_str='hello')
-
-    results = bench.get_results()
-    args = results['args'][0]
-    kwargs = results['kwargs'][0]
-
-    # Values must be their native Python types, not JSON-encoded strings
-    assert args[0] == 42, f'Expected int 42, got {args[0]!r}'
-    assert args[1] == {'key': 'value'}, f'Expected dict, got {args[1]!r}'
-    assert kwargs['kw_str'] == 'hello', f'Expected str hello, got {kwargs["kw_str"]!r}'
-
-
-def test_unjsonencodable_arg_kwarg_retval():
-    class Bench(MicroBench, MBFunctionCall, MBReturnValue):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def dummy(arg1, arg2):
-        return object()
-
-    with warnings.catch_warnings(record=True) as w:
-        # Run a function with unencodable arg, kwarg, return value
-        dummy(object(), arg2=object())
-
-        # Check that we get three warnings - one each for args,
-        # kwargs, return value
-        assert len(w) == 3
-        assert all(issubclass(w_.category, JSONEncodeWarning) for w_ in w)
-
-    results = bench.get_results()
-    assert results['args'][0] == [_UNENCODABLE_PLACEHOLDER_VALUE]
-    assert results['kwargs'][0] == {'arg2': _UNENCODABLE_PLACEHOLDER_VALUE}
-    assert results['return_value'][0] == _UNENCODABLE_PLACEHOLDER_VALUE
-
-
-def test_custom_jsonencoder():
-    # A custom class which can't be encoded to JSON by default
-    class MyCustomClass:
-        def __init__(self, message):
-            self.message = message
-
-        def __str__(self):
-            return f'<MyCustomClass "{self.message}">'
-
-    # Implement JSON encoding for objects of the above class
-    class CustomJSONEncoder(JSONEncoder):
-        def default(self, o):
-            if isinstance(o, MyCustomClass):
-                return str(o)
-
-            return super().default(o)
-
-    class Bench(MicroBench, MBReturnValue):
-        pass
-
-    # Create a benchmark suite with custom JSON encoder
-    bench = Bench(json_encoder=CustomJSONEncoder)
-
-    # Custom object which requires special handling for JSONEncoder
-    obj = MyCustomClass('test message')
-
-    @bench
-    def dummy():
-        return obj
-
-    dummy()
-
-    results = bench.get_results()
-    assert results['return_value'][0] == str(obj)
-
-
-def test_jsonencoder_numpy_types():
-    """JSONEncoder handles numpy integer, float, and ndarray natively."""
-    encoder = JSONEncoder()
-    assert encoder.default(numpy.int64(7)) == 7
-    assert encoder.default(numpy.float32(3.14)) == pytest.approx(3.14, abs=1e-5)
-    assert encoder.default(numpy.array([1, 2, 3])) == [1, 2, 3]
-
-
-def test_jsonencoder_timedelta_and_timezone():
-    """JSONEncoder serialises timedelta as total seconds and timezone as string."""
-    encoder = JSONEncoder()
-    assert encoder.default(datetime.timedelta(seconds=90)) == 90.0
-    tz = datetime.timezone(datetime.timedelta(hours=5, minutes=30))
-    tz_str = encoder.default(tz)
-    assert '05:30' in tz_str
-
-
-def test_positional_args_raises():
-    """MicroBench constructor rejects extra positional arguments.
-
-    The *args guard is primarily designed for subclasses that forward *args
-    via super().__init__(*args, **kwargs). Triggering it directly requires
-    saturating the seven named positional parameters first.
-    """
-    with pytest.raises(ValueError, match='keyword'):
-        MicroBench(None, JSONEncoder, datetime.timezone.utc, 1, 0, None, None, 'extra')
-
-
-def test_outfile_and_outputs_raises():
-    """Passing both outfile and outputs raises ValueError."""
-    with pytest.raises(ValueError, match='mutually exclusive'):
-        MicroBench(outfile='/tmp/x.json', outputs=[FileOutput()])
-
-
-def test_outfile_string_path():
-    """Results are written to and read from a file-system path."""
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.json', delete=False) as f:
-        tmppath = f.name
-
-    try:
-        bench = MicroBench(outfile=tmppath)
-
-        @bench
-        def noop():
-            pass
-
-        noop()
-
-        assert os.path.getsize(tmppath) > 0
-        results = pandas.read_json(tmppath, lines=True)
-        assert results['function_name'][0] == 'noop'
-    finally:
-        os.unlink(tmppath)
-
-
-def test_env_vars_not_iterable():
-    """env_vars must be iterable; a non-iterable raises ValueError."""
-
-    class BadBench(MicroBench):
-        env_vars = 42  # not iterable
-
-    bench = BadBench()
-
-    @bench
-    def noop():
-        pass
-
-    with pytest.raises(ValueError, match='env_vars'):
-        noop()
-
-
-def test_capture_versions_not_iterable():
-    """capture_versions must be iterable; a non-iterable raises ValueError."""
-
-    class BadBench(MicroBench):
-        capture_versions = 42  # not iterable
-
-    bench = BadBench()
-
-    @bench
-    def noop():
-        pass
-
-    with pytest.raises(ValueError, match='capture_versions'):
-        noop()
-
-
-def test_get_results_without_pandas():
-    """get_results raises ImportError when pandas is unavailable."""
-    import microbench
-
-    bench = MicroBench()
-
-    with patch.object(microbench, 'pandas', None):
-        with pytest.raises(ImportError, match='pandas'):
-            bench.get_results()
-
-
-def test_monitor_multiple_samples():
-    """MonitorThread collects more than one sample for a long-running function."""
-
-    class MonitorBench(MicroBench):
-        monitor_interval = 0.05
-
-        @staticmethod
-        def monitor(process):
-            return {'rss': process.memory_info().rss}
-
-    monitor_bench = MonitorBench()
-
-    @monitor_bench
-    def slow_function():
-        time.sleep(0.25)
-
-    slow_function()
-
-    results = monitor_bench.get_results()
-    assert len(results['monitor'][0]) >= 2, 'Expected at least 2 monitor samples'
-
-
-def test_multi_sink_output():
-    """Results are written to all configured output sinks."""
-
-    class RecordingOutput(Output):
-        def __init__(self):
-            self.records = []
-
-        def write(self, bm_json_str):
-            self.records.append(bm_json_str)
-
-    sink_a = RecordingOutput()
-    sink_b = FileOutput()
-
-    bench = MicroBench(outputs=[sink_a, sink_b])
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-    noop()
-
-    assert len(sink_a.records) == 2
-    results = sink_b.get_results()
-    assert len(results) == 2
-    assert (results['function_name'] == 'noop').all()
-
-
-def test_output_base_get_results_raises():
-    """Output.get_results() raises NotImplementedError by default."""
-    with pytest.raises(NotImplementedError):
-        Output().get_results()
-
-
-def test_no_supporting_sink_raises():
-    """get_results() raises RuntimeError when no sink supports it."""
-
-    class SinkOnly(Output):
-        def write(self, bm_json_str):
-            pass
-
-    bench = MicroBench(outputs=[SinkOnly()])
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-
-    with pytest.raises(RuntimeError, match='get_results'):
-        bench.get_results()
-
-
-def _make_mock_redis(redis_store):
-    """Return a mock redis module wired to the given list as a backing store."""
-    mock_redis_client = MagicMock()
-    mock_redis_client.rpush.side_effect = lambda key, val: redis_store.append(
-        val.encode('utf8') if isinstance(val, str) else val
-    )
-    mock_redis_client.lrange.side_effect = lambda key, start, end: redis_store
-    mock_redis = MagicMock()
-    mock_redis.StrictRedis.return_value = mock_redis_client
-    return mock_redis, mock_redis_client
-
-
-def test_redis_output_get_results():
-    """RedisOutput.get_results() reads results back from Redis."""
-    redis_store = []
-    mock_redis, mock_redis_client = _make_mock_redis(redis_store)
-
-    with patch.dict('sys.modules', {'redis': mock_redis}):
-        bench = MicroBench(outputs=[RedisOutput('test:bench')])
-
-        @bench
-        def noop():
-            pass
-
-        noop()
-
-        results = bench.get_results()
-        assert results['function_name'][0] == 'noop'
-        assert 'start_time' in results.columns
-        assert 'finish_time' in results.columns
-
-        mock_redis_client.rpush.assert_called_once()
-        assert mock_redis_client.rpush.call_args[0][0] == 'test:bench'
-
-
-def test_redis_output_get_results_without_pandas():
-    """RedisOutput.get_results() raises ImportError without pandas."""
-    import microbench
-
-    redis_store = []
-    mock_redis, _ = _make_mock_redis(redis_store)
-
-    with patch.dict('sys.modules', {'redis': mock_redis}):
-        bench = MicroBench(outputs=[RedisOutput('test:bench')])
-
-        with patch.object(microbench, 'pandas', None):
-            with pytest.raises(ImportError, match='pandas'):
-                bench.get_results()
-
-
-def test_mb_peak_memory():
-    class Bench(MicroBench, MBPeakMemory):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def allocate():
-        return [0] * 100_000
-
-    allocate()
-
-    results = bench.get_results()
-    assert results['peak_memory_bytes'][0] > 0
-
-
-def test_mb_peak_memory_stops_tracemalloc():
-    """MBPeakMemory stops tracemalloc after the call if it was not already running."""
-    import tracemalloc
-
-    assert not tracemalloc.is_tracing()
-
-    class Bench(MicroBench, MBPeakMemory):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-
-    assert not tracemalloc.is_tracing()
-
-
-def test_mb_peak_memory_preserves_existing_trace():
-    """MBPeakMemory leaves tracemalloc running if it was already active."""
-    import tracemalloc
-
-    tracemalloc.start()
-    try:
-
-        class Bench(MicroBench, MBPeakMemory):
-            pass
-
-        bench = Bench()
-
-        @bench
-        def noop():
-            pass
-
-        noop()
-
-        assert tracemalloc.is_tracing()
-        results = bench.get_results()
-        assert 'peak_memory_bytes' in results.columns
-    finally:
-        tracemalloc.stop()
-
-
-def test_redis_output_multiple_results():
-    """RedisOutput.get_results() returns all stored results."""
-    redis_store = []
-    mock_redis, _ = _make_mock_redis(redis_store)
-
-    with patch.dict('sys.modules', {'redis': mock_redis}):
-        bench = MicroBench(outputs=[RedisOutput('test:bench')])
-
-        @bench
-        def func_a():
-            pass
-
-        @bench
-        def func_b():
-            pass
-
-        func_a()
-        func_b()
-
-        results = bench.get_results()
-        assert len(results) == 2
-        assert list(results['function_name']) == ['func_a', 'func_b']
-
-
-# ---------------------------------------------------------------------------
-# MBGitInfo
-# ---------------------------------------------------------------------------
-
-_GIT_TOPLEVEL = b'/home/user/project\n'
-_GIT_STATUS_CLEAN = (
-    '# branch.oid abc123def456abc123def456abc123def456abc1\n# branch.head main\n'
-)
-_GIT_STATUS_DIRTY = _GIT_STATUS_CLEAN + ' M modified_file.py\n'
-_GIT_STATUS_DETACHED = (
-    '# branch.oid abc123def456abc123def456abc123def456abc1\n# branch.head (detached)\n'
-)
-
-
-def _git_mock(status_output):
-    """Return a subprocess.check_output side_effect that handles both git calls."""
-
-    def side_effect(cmd, **kwargs):
-        if 'rev-parse' in cmd:
-            return _GIT_TOPLEVEL
-        return status_output.encode()
-
-    return side_effect
-
-
-def test_mb_git_info():
-    """MBGitInfo captures repo, commit, branch, and dirty=False from a clean repo."""
-
-    class Bench(MicroBench, MBGitInfo):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch('subprocess.check_output', side_effect=_git_mock(_GIT_STATUS_CLEAN)):
-        noop()
-
-    git_info = bench.get_results()['git_info'][0]
-    assert git_info['repo'] == '/home/user/project'
-    assert git_info['commit'] == 'abc123def456abc123def456abc123def456abc1'
-    assert git_info['branch'] == 'main'
-    assert git_info['dirty'] is False
-
-
-def test_mb_git_info_dirty():
-    """MBGitInfo sets dirty=True when the working tree has uncommitted changes."""
-
-    class Bench(MicroBench, MBGitInfo):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch('subprocess.check_output', side_effect=_git_mock(_GIT_STATUS_DIRTY)):
-        noop()
-
-    assert bench.get_results()['git_info'][0]['dirty'] is True
-
-
-def test_mb_git_info_detached_head():
-    """MBGitInfo stores an empty string for branch in detached HEAD state."""
-
-    class Bench(MicroBench, MBGitInfo):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch('subprocess.check_output', side_effect=_git_mock(_GIT_STATUS_DETACHED)):
-        noop()
-
-    assert bench.get_results()['git_info'][0]['branch'] == ''
-
-
-def test_mb_git_info_no_git_raises():
-    """MBGitInfo propagates FileNotFoundError when git is not on PATH."""
-
-    class Bench(MicroBench, MBGitInfo):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch(
-        'subprocess.check_output', side_effect=FileNotFoundError('git not found')
-    ):
-        with pytest.raises(FileNotFoundError):
-            noop()
-
-
-def test_mb_git_info_default_uses_script_dir():
-    """MBGitInfo defaults to the directory of sys.argv[0], not CWD."""
-    captured_kwargs = []
-
-    def side_effect(cmd, **kwargs):
-        captured_kwargs.append(kwargs)
-        return _GIT_TOPLEVEL if 'rev-parse' in cmd else _GIT_STATUS_CLEAN.encode()
-
-    class Bench(MicroBench, MBGitInfo):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    script = '/some/project/script.py'
-    expected_cwd = os.path.dirname(os.path.abspath(script))
-
-    with patch.object(sys, 'argv', [script]):
-        with patch('subprocess.check_output', side_effect=side_effect):
-            noop()
-
-    assert all(kw.get('cwd') == expected_cwd for kw in captured_kwargs)
-
-
-def test_mb_git_info_custom_path():
-    """MBGitInfo passes git_repo as cwd to subprocess."""
-    captured_kwargs = []
-
-    def side_effect(cmd, **kwargs):
-        captured_kwargs.append(kwargs)
-        return _GIT_TOPLEVEL if 'rev-parse' in cmd else _GIT_STATUS_CLEAN.encode()
-
-    class Bench(MicroBench, MBGitInfo):
-        git_repo = '/some/repo'
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch('subprocess.check_output', side_effect=side_effect):
-        noop()
-
-    assert all(kw.get('cwd') == '/some/repo' for kw in captured_kwargs)
-
-
-# ---------------------------------------------------------------------------
-# MBFileHash
-# ---------------------------------------------------------------------------
-
-
-def test_mb_file_hash_specific_file(tmp_path):
-    """MBFileHash records the SHA-256 digest of a specified file."""
-    import hashlib
-
-    content = b'hello microbench'
-    target = tmp_path / 'data.txt'
-    target.write_bytes(content)
-
-    expected = hashlib.sha256(content).hexdigest()
-
-    class Bench(MicroBench, MBFileHash):
-        hash_files = [str(target)]
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-
-    results = bench.get_results()
-    assert results['file_hashes'][0] == {str(target): expected}
-
-
-def test_mb_file_hash_multiple_files(tmp_path):
-    """MBFileHash records digests for all specified files."""
-    import hashlib
-
-    files = {}
-    for name in ('a.txt', 'b.txt'):
-        p = tmp_path / name
-        p.write_bytes(name.encode())
-        files[str(p)] = hashlib.sha256(name.encode()).hexdigest()
-
-    class Bench(MicroBench, MBFileHash):
-        hash_files = list(files.keys())
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-
-    results = bench.get_results()
-    assert results['file_hashes'][0] == files
-
-
-def test_mb_file_hash_custom_algorithm(tmp_path):
-    """MBFileHash respects the hash_algorithm class attribute."""
-    import hashlib
-
-    content = b'test data'
-    target = tmp_path / 'script.py'
-    target.write_bytes(content)
-
-    expected = hashlib.md5(content).hexdigest()
-
-    class Bench(MicroBench, MBFileHash):
-        hash_files = [str(target)]
-        hash_algorithm = 'md5'
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    noop()
-
-    results = bench.get_results()
-    assert results['file_hashes'][0] == {str(target): expected}
-
-
-def test_mb_file_hash_default_uses_argv(tmp_path):
-    """MBFileHash defaults to sys.argv[0] when hash_files is not set."""
-    import hashlib
-
-    content = b'print("hello")'
-    script = tmp_path / 'run.py'
-    script.write_bytes(content)
-
-    expected = hashlib.sha256(content).hexdigest()
-
-    class Bench(MicroBench, MBFileHash):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch.object(sys, 'argv', [str(script)]):
-        noop()
-
-    results = bench.get_results()
-    assert results['file_hashes'][0] == {str(script): expected}
-
-
-def test_mb_file_hash_no_argv_empty(tmp_path):
-    """MBFileHash records an empty dict when sys.argv is empty."""
-
-    class Bench(MicroBench, MBFileHash):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch.object(sys, 'argv', []):
-        noop()
-
-    results = bench.get_results()
-    assert results['file_hashes'][0] == {}
-
-
-def test_mb_file_hash_interactive_argv_empty_string(tmp_path):
-    """MBFileHash records an empty dict in interactive Python (argv[0] == '')."""
-
-    class Bench(MicroBench, MBFileHash):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with patch.object(sys, 'argv', ['']):
-        noop()
-
-    results = bench.get_results()
-    assert results['file_hashes'][0] == {}
-
-
-def test_mb_file_hash_missing_file_raises(tmp_path):
-    """MBFileHash raises FileNotFoundError for a path that does not exist."""
-
-    class Bench(MicroBench, MBFileHash):
-        hash_files = [str(tmp_path / 'nonexistent.py')]
-
-    bench = Bench()
-
-    @bench
-    def noop():
-        pass
-
-    with pytest.raises(FileNotFoundError):
-        noop()
-
-
-# ---------------------------------------------------------------------------
-# bench.record() context manager
-# ---------------------------------------------------------------------------
-
-
-def test_record_standard_fields():
-    """bench.record() produces a record with all standard timing fields."""
-    bench = MicroBench()
-
-    with bench.record('my_block'):
-        pass
-
-    results = bench.get_results()
-    assert len(results) == 1
-    row = results.iloc[0]
-    assert row['function_name'] == 'my_block'
-    assert 'start_time' in results.columns
-    assert 'finish_time' in results.columns
-    assert len(row['run_durations']) == 1
-    assert 'mb_run_id' in results.columns
-    assert 'mb_version' in results.columns
-
-
-def test_record_no_name_defaults():
-    """bench.record() with no name sets function_name to '<record>'."""
-    bench = MicroBench()
-
-    with bench.record():
-        pass
-
-    results = bench.get_results()
-    assert results.iloc[0]['function_name'] == '<record>'
-
-
-def test_record_static_fields():
-    """Static fields passed to MicroBench() appear in bench.record() output."""
-    bench = MicroBench(experiment='run-1', trial=3)
-
-    with bench.record('block'):
-        pass
-
-    results = bench.get_results()
-    assert results.iloc[0]['experiment'] == 'run-1'
-    assert results.iloc[0]['trial'] == 3
-
-
-def test_record_mixin_fields():
-    """Mixin capture methods run during bench.record()."""
-
-    class Bench(MicroBench, MBHostInfo):
-        pass
-
-    bench = Bench()
-
-    with bench.record('block'):
-        pass
-
-    results = bench.get_results()
-    assert 'hostname' in results.columns
-    assert 'operating_system' in results.columns
-
-
-def test_record_multiple_records():
-    """Each bench.record() call appends a separate record."""
-    bench = MicroBench()
-
-    with bench.record('first'):
-        pass
-    with bench.record('second'):
-        pass
-
-    results = bench.get_results()
-    assert len(results) == 2
-    assert list(results['function_name']) == ['first', 'second']
-
-
-def test_record_exception_captured_and_reraised():
-    """Exceptions inside bench.record() are recorded then re-raised."""
-    bench = MicroBench()
-
-    with pytest.raises(ValueError, match='oops'):
-        with bench.record('block'):
-            raise ValueError('oops')
-
-    results = bench.get_results()
-    assert len(results) == 1
-    exc = results.iloc[0]['exception']
-    assert exc['type'] == 'ValueError'
-    assert exc['message'] == 'oops'
-
-
-def test_record_no_exception_field_on_success():
-    """No 'exception' field is present when the block completes normally."""
-    bench = MicroBench()
-
-    with bench.record('block'):
-        pass
-
-    results = bench.get_results()
-    assert 'exception' not in results.columns
-
-
-def test_record_coexists_with_decorator():
-    """bench.record() and @bench decorator write to the same output sink."""
-    bench = MicroBench()
-
-    with bench.record('ctx'):
-        pass
-
-    @bench
-    def decorated():
-        pass
-
-    decorated()
-
-    results = bench.get_results()
-    assert len(results) == 2
-    assert set(results['function_name']) == {'ctx', 'decorated'}
-
-
-# ---------------------------------------------------------------------------
-# Exception capture — decorator
-# ---------------------------------------------------------------------------
-
-
-def test_decorator_exception_captured_and_reraised():
-    """Exceptions from @bench functions are recorded then re-raised."""
-    bench = MicroBench()
-
-    @bench
-    def failing():
-        raise RuntimeError('boom')
-
-    with pytest.raises(RuntimeError, match='boom'):
-        failing()
-
-    results = bench.get_results()
-    assert len(results) == 1
-    exc = results.iloc[0]['exception']
-    assert exc['type'] == 'RuntimeError'
-    assert exc['message'] == 'boom'
-
-
-def test_decorator_no_exception_field_on_success():
-    """No 'exception' field when the decorated function returns normally."""
-    bench = MicroBench()
-
-    @bench
-    def ok():
-        pass
-
-    ok()
-
-    results = bench.get_results()
-    assert 'exception' not in results.columns
-
-
-def test_decorator_exception_stops_iterations():
-    """An exception on iteration 2 stops the loop; only 2 durations recorded."""
-    call_count = 0
-
-    bench = MicroBench(iterations=5)
-
-    @bench
-    def sometimes_fails():
-        nonlocal call_count
-        call_count += 1
-        if call_count == 2:
-            raise ValueError('fail on second')
-
-    with pytest.raises(ValueError):
-        sometimes_fails()
-
-    results = bench.get_results()
-    assert call_count == 2
-    assert len(results.iloc[0]['run_durations']) == 2
-
-
-def test_decorator_return_value_mixin_skipped_on_exception():
-    """MBReturnValue does not set return_value when the function raised."""
-
-    class Bench(MicroBench, MBReturnValue):
-        pass
-
-    bench = Bench()
-
-    @bench
-    def failing():
-        raise TypeError('no return')
-
-    with pytest.raises(TypeError):
-        failing()
-
-    results = bench.get_results()
-    assert 'return_value' not in results.columns
-
-
-# ---------------------------------------------------------------------------
-# Mixin behaviour with bench.record()
-# ---------------------------------------------------------------------------
-
-
-def test_record_mbfunctioncall_produces_empty_args():
-    """MBFunctionCall with bench.record() records args=[] kwargs={} (no callable)."""
-
-    class Bench(MicroBench, MBFunctionCall):
-        pass
-
-    bench = Bench()
-
-    with bench.record('block'):
-        pass
-
-    results = bench.get_results()
-    assert results.iloc[0]['args'] == []
-    assert results.iloc[0]['kwargs'] == {}
-
-
-def test_record_mbreturnvalue_ignored():
-    """MBReturnValue is silently a no-op with bench.record() (no return value)."""
-
-    class Bench(MicroBench, MBReturnValue):
-        pass
-
-    bench = Bench()
-
-    with bench.record('block'):
-        pass
-
-    results = bench.get_results()
-    assert 'return_value' not in results.columns
-
-
-def test_record_mblineprofiler_raises():
-    """MBLineProfiler raises NotImplementedError with bench.record()."""
-    from microbench import MBLineProfiler
-
-    class Bench(MicroBench, MBLineProfiler):
-        pass
-
-    bench = Bench()
-
-    with pytest.raises(NotImplementedError, match='MBLineProfiler'):
-        with bench.record('block'):
-            pass
-
-
-# ---------------------------------------------------------------------------
-# bench.record_on_exit()
-# ---------------------------------------------------------------------------
-
-
-def _invoke_record_on_exit(bench):
-    """Directly call the registered exit handler and return its results."""
-    bench._record_on_exit_handler()
-    return bench.get_results()
-
-
-def test_record_on_exit_standard_fields():
-    """record_on_exit() produces a record with all standard timing fields."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit('simulation')
-        results = _invoke_record_on_exit(bench)
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert len(results) == 1
-    row = results.iloc[0]
-    assert row['function_name'] == 'simulation'
-    assert len(row['run_durations']) == 1
-    assert row['run_durations'][0] >= 0
-    assert 'start_time' in results.columns
-    assert 'finish_time' in results.columns
-    assert 'mb_run_id' in results.columns
-    assert 'mb_version' in results.columns
-
-
-def test_record_on_exit_default_name():
-    """record_on_exit() with no name sets function_name to '<process>'."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit()
-        results = _invoke_record_on_exit(bench)
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert results.iloc[0]['function_name'] == '<process>'
-
-
-def test_record_on_exit_static_fields():
-    """Static fields from MicroBench() constructor appear in the record."""
-    bench = MicroBench(experiment='run-1', trial=7)
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit('sim')
-        results = _invoke_record_on_exit(bench)
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert results.iloc[0]['experiment'] == 'run-1'
-    assert results.iloc[0]['trial'] == 7
-
-
-def test_record_on_exit_mixin_fields():
-    """Mixin capture methods run in the exit handler."""
-
-    class Bench(MicroBench, MBHostInfo):
-        pass
-
-    bench = Bench()
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit('sim')
-        results = _invoke_record_on_exit(bench)
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert 'hostname' in results.columns
-    assert 'operating_system' in results.columns
-
-
-def test_record_on_exit_exception_capture():
-    """Unhandled exceptions are captured via sys.excepthook."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit('sim')
-        # Simulate Python calling sys.excepthook for an unhandled exception.
-        exc = RuntimeError('something broke')
-        sys.excepthook(type(exc), exc, None)
-        results = _invoke_record_on_exit(bench)
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert len(results) == 1
-    exc_field = results.iloc[0]['exception']
-    assert exc_field['type'] == 'RuntimeError'
-    assert exc_field['message'] == 'something broke'
-
-
-def test_record_on_exit_no_exception_field_on_clean_exit():
-    """No 'exception' field when the process exits cleanly."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit('sim')
-        results = _invoke_record_on_exit(bench)
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert 'exception' not in results.columns
-
-
-def test_record_on_exit_sigterm_writes_record():
-    """The SIGTERM handler writes the record with exit_signal='SIGTERM'."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    orig_sigterm = _signal.getsignal(_signal.SIGTERM)
-    try:
-        bench.record_on_exit('sim', handle_sigterm=True)
-        handler = _signal.getsignal(_signal.SIGTERM)
-        assert callable(handler)
-        # Invoke the handler but prevent it from re-killing the process.
-        with patch('os.kill'), patch('signal.signal'):
-            handler(_signal.SIGTERM, None)
-        results = bench.get_results()
-    finally:
-        sys.excepthook = orig_excepthook
-        _signal.signal(_signal.SIGTERM, orig_sigterm)
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert len(results) == 1
-    assert results.iloc[0]['exit_signal'] == 'SIGTERM'
-
-
-def test_record_on_exit_handle_sigterm_false():
-    """handle_sigterm=False leaves the SIGTERM handler unchanged."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    orig_sigterm = _signal.getsignal(_signal.SIGTERM)
-    try:
-        bench.record_on_exit('sim', handle_sigterm=False)
-        assert _signal.getsignal(_signal.SIGTERM) is orig_sigterm
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-
-def test_record_on_exit_double_fire_prevention():
-    """Calling the exit handler twice writes only one record."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit('sim')
-        bench._record_on_exit_handler()
-        bench._record_on_exit_handler()
-        results = bench.get_results()
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert len(results) == 1
-
-
-def test_record_on_exit_re_registration_replaces_first():
-    """A second record_on_exit() call replaces the first registration."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit('first')
-        first_handler = bench._record_on_exit_handler
-        bench.record_on_exit('second')
-        second_handler = bench._record_on_exit_handler
-
-        # Old handler should be deregistered; calling it is now a no-op
-        # because it is no longer in the atexit list — but it would still
-        # run if called directly. The key check is that they are distinct.
-        assert first_handler is not second_handler
-
-        second_handler()
-        results = bench.get_results()
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    assert len(results) == 1
-    assert results.iloc[0]['function_name'] == 'second'
-
-
-def test_record_on_exit_non_main_thread_warns():
-    """record_on_exit() warns when called from a non-main thread."""
-    bench = MicroBench()
-    orig_excepthook = sys.excepthook
-    orig_sigterm = _signal.getsignal(_signal.SIGTERM)
-    warning_holder = []
-
-    def run():
-        with warnings.catch_warnings(record=True) as w:
-            warnings.simplefilter('always')
-            bench.record_on_exit('sim', handle_sigterm=True)
-            warning_holder.extend(w)
-
-    t = threading.Thread(target=run)
-    t.start()
-    t.join()
-    try:
-        assert any(issubclass(w.category, RuntimeWarning) for w in warning_holder)
-        assert _signal.getsignal(_signal.SIGTERM) is orig_sigterm
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-
-def test_record_on_exit_output_fallback_to_stderr(capsys):
-    """When output_result raises, the record is written to stderr."""
-
-    class BrokenOutput(Output):
-        def write(self, bm_json_str):
-            raise OSError('disk full')
-
-    bench = MicroBench(outputs=[BrokenOutput()])
-    orig_excepthook = sys.excepthook
-    try:
-        bench.record_on_exit('sim')
-        bench._record_on_exit_handler()
-    finally:
-        sys.excepthook = orig_excepthook
-        _atexit.unregister(bench._record_on_exit_handler)
-
-    captured = capsys.readouterr()
-    import json as _json
-
-    record = _json.loads(captured.err.strip())
-    assert record['function_name'] == 'sim'
